% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion kernels 150419.R
\name{kernel_sRD}
\alias{kernel_sRD}
\title{Kernel function}
\usage{
kernel_sRD(kernel_, weights, h, X, c)
}
\arguments{
\item{kernel_}{kernel type: triangular, epanechnikov, tricube, uniform, gaussian y cosine}

\item{h}{bandwidth}

\item{X}{continuous variable}

\item{c}{cutpoint}

\item{weigths}{pik, NULL if there aren't sample weights}
}
\value{
matriz con pesos kernel incluyendo o no los pesos de muestreo
}
\description{
estimation of the kernel function using sample weights
}
\examples{

}
