% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion parametrica 140419.R
\name{RDpar}
\alias{RDpar}
\title{RDD Parametric}
\usage{
RDpar(D = muestra1$D, Z = muestra1$x, Y = muestra1$y, c = 0,
  type = "sharp", TreatLeft = T, W = NULL, svydesign_ = muestra,
  vcov_)
}
\arguments{
\item{D}{variable with: 1 for treatment group and 0 for control group}

\item{Z}{continuous variable}

\item{Y}{outcome variable}

\item{c}{cutpoint}

\item{type}{sharp or fuzzy}

\item{TreatLeft}{T if the treatment is applied for below of the cutpoint}

\item{W}{sample weights 1/pik, NULL if there aren't sample weights}

\item{svydesign_}{design sample using survey library}

\item{vcov_}{variances and covariances type: HCO, HC1, HC2 y HC3}
}
\value{
Impact estimation, e.e. and t test
}
\description{
parametric regression discontinuity using the likelihood and pseudolikelihood estimator with and without sample weights
}
\details{
RDD PARAMETRIC
}
\examples{

}
