% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion polinomios 150419.R
\name{RDpol}
\alias{RDpol}
\title{RDD poliynomial local regression}
\usage{
RDpol(D = muestra1$D, Z = muestra1$x, Y = muestra1$y, c = 0,
  type = "sharp", TreatLeft = T, weights = NULL,
  svydesign_ = muestra, vcov_, kernel_ = "uniforme", h = 2, q = 2)
}
\arguments{
\item{D}{variable with: 1 for treatment group and 0 for control group}

\item{Z}{continuous variable}

\item{Y}{outcome variable}

\item{c}{cutpoint}

\item{type}{sharp or fuzzy}

\item{TreatLeft}{T if the treatment is applied for below of the cutpoint}

\item{svydesign_}{design sample using survey library}

\item{vcov_}{variances and covariances type: HCO, HC1, HC2 y HC3}

\item{kernel_}{kernel type: triangular, epanechnikov, tricube, uniform, gaussian y cosine}

\item{h}{bandwidth}

\item{q}{order polynomial}

\item{W}{sample weights 1/pik, NULL if there aren't sample weights}
}
\value{
Impact estimation, e.e. and t test
}
\description{
no parametric regression discontinuity using polynomial local regression estimator with and without sample weights
}
\examples{

}
